Déplacer enregistrement du service dialog
